Drop RowModel.setDatabaseValue(_:forColumn:).
This leaves RowModel.updateFromRow(_) as the only method that
updates a RowModel. RowModel.updateFromRow was introduced in order to let models
update compound properties such as CLLocationCoordinate2D. The problem was that
RowModel.updateFromRow could be overriden so that it would handle 100% of the
updating job, leaving RowModel.setDatabaseValue as a method that could not
reliably be used to update a RowModel, and thus had to be dropped.