Handle null primary keys in PersistableRecord #1023
If a PersistableRecord has a multi-column primary key and some of those columns are nil then will now generate IS NULL sql for those nil columns.

Note, it remains the case that if all the columns involved in a primary key are null it will exit early, without querying the database.