Fix quoting of FTS5 tokenizer arguments
Tokenizer arguments only need to be quoted ("foo" -> "'foo'") when we generate SQL for a tokenizer. They must not be quoted when we instantiate a tokenizer with the fts5_tokenizer.xCreate function.

This commit fixes tokenizer instantiation by not quoting arguments when we call the fts5_tokenizer.xCreate function.

Before this fix, the FTS5Tokenizer.tokenize(query:) method would return incorrect tokens.